main-server/requirements.txt
fastapi
pymongo
uvicorn[standard]
grpcio==1.47.0
grpcio-tools==1.47.0